+++ engine/src/engine/search/negamax.rs
@@ -154,12 +154,12 @@ pub fn negamax(
             tt_entry.eval
         };
 
-        let eval = (raw_eval + ctx.tables.corrhist.get(game)).clamp_to_non_mate();
+        let eval = (raw_eval + ctx.tables.corrhist.get(game, ctx, plies)).clamp_to_non_mate();
 
         (raw_eval, eval)
     } else {
         let raw_eval = eval::eval(ctx.nnue, game);
-        let eval = (raw_eval + ctx.tables.corrhist.get(game)).clamp_to_non_mate();
+        let eval = (raw_eval + ctx.tables.corrhist.get(game, ctx, plies)).clamp_to_non_mate();
 
         ctx.tt.insert(
             game.hash,
@@ -578,7 +578,9 @@ pub fn negamax(
             || tt_node_bound == NodeBound::Lower && best_score <= eval
             || tt_node_bound == NodeBound::Upper && best_score >= eval)
         {
-            ctx.tables.corrhist.update(game, depth, best_score - eval);
+            ctx.tables
+                .corrhist
+                .update(game, ctx.stack, plies, depth, best_score - eval);
         }
 
         ctx.tt.insert(